requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
scipy>=1.10.0
scikit-image>=0.21.0
pytest>=7.4.0
black>=23.0.0
mypy>=1.6.0
pre-commit>=3.5.0